+++ lang_qc/util/qc_state_helpers.py
@@ -1,0 +1,187 @@
+# Copyright (c) 2022 Genome Research Ltd.
+#
+# Author: Adam Blanchet <ab59@sanger.ac.uk>
+#
+# This file is part of npg_langqc.
+#
+# npg_langqc is free software: you can redistribute it and/or modify it under
+# the terms of the GNU General Public License as published by the Free Software
+# Foundation; either version 3 of the License, or (at your option) any later
+# version.
+#
+# This program is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+# FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
+# details.
+#
+# You should have received a copy of the GNU General Public License along with
+# this program. If not, see <http://www.gnu.org/licenses/>.
+import json
+from datetime import datetime
+from typing import Optional
+
+from product_id.main import PacBioWell
+from sqlalchemy.orm import Session
+from sqlalchemy import select, and_
+
+from lang_qc.db.qc_schema import (
+    ProductLayout,
+    QcState,
+    QcStateDict,
+    QcType,
+    SeqPlatform,
+    SeqProduct,
+    SubProduct,
+    SubProductAttr,
+    User,
+)
+from lang_qc.models.inbox_models import QcStatus
+
+
+def create_id_product(run_name, well_label):
+    return PacBioWell(run_name=run_name, well_label=well_label)
+
+
+def create_well_properties(run_name, well_label):
+    return json.dumps({"run_name": run_name, "well_label": well_label})
+
+
+def create_well_properties_digest(run_name, well_label):
+    return hash(frozenset(create_well_properties(run_name, well_label)))
+
+
+def get_seq_product_for_well(run_name: str, well_label: str, qcdb_session: Session):
+    """Get a SeqProduct from a well."""
+
+    return (
+        qcdb_session.execute(
+            select(SeqProduct)
+            .join(ProductLayout)
+            .join(SubProduct)
+            .where(
+                and_(
+                    SubProduct.value_attr_one == run_name,
+                    SubProduct.value_attr_two == well_label,
+                )
+            )
+        )
+        .scalars()
+        .one_or_none()
+    )
+
+
+def get_qc_state_for_well(
+    run_name: str, well_label: str, qcdb_session: Session
+) -> Optional[QcState]:
+    """Get a QcState from a well."""
+
+    return qcdb_session.execute(
+        select(QcState)
+        .join(SeqProduct)
+        .join(ProductLayout)
+        .join(SubProduct)
+        .where(
+            and_(
+                SubProduct.value_attr_one == run_name,
+                SubProduct.value_attr_two == well_label,
+            )
+        )
+    ).scalar_one_or_none()
+
+
+def construct_seq_product_for_well(
+    run_name: str, well_label: str, qcdb_session: Session
+):
+
+    seq_platform = qcdb_session.execute(
+        select(SeqPlatform).where(SeqPlatform.name == "PacBio")
+    ).scalar_one_or_none()
+    if seq_platform is None:
+        raise Exception("PacBio SeqPlatform is not in the QC database.")
+
+    run_name_product_attr = qcdb_session.execute(
+        select(SubProductAttr).where(SubProductAttr.attr_name == "run_name")
+    ).scalar_one_or_none()
+    if run_name_product_attr is None:
+        raise Exception("PacBio run name SubProductAttr is not the QC database.")
+
+    well_label_product_attr = qcdb_session.execute(
+        select(SubProductAttr).where(SubProductAttr.attr_name == "well_label")
+    ).scalar_one_or_none()
+    if well_label_product_attr is None:
+        raise Exception("PacBio well label SubProductAttr is not in the QC database.")
+
+    seq_product = SeqProduct(
+        id_product=create_id_product(run_name, well_label),
+        seq_platform=seq_platform,
+        product_layout=[
+            ProductLayout(
+                sub_product=SubProduct(
+                    sub_product_attr=run_name_product_attr,
+                    sub_product_attr_=well_label_product_attr,
+                    value_attr_one=run_name,
+                    value_attr_two=well_label,
+                    properties=create_well_properties(run_name, well_label),
+                    properties_digest=create_well_properties_digest(
+                        run_name, well_label
+                    ),
+                )
+            )
+        ],
+    )
+
+    qcdb_session.add(seq_product)
+    qcdb_session.commit()
+
+    return seq_product
+
+
+def update_qc_state(
+    qc_status_post: QcStatus, qc_state_db: QcState, qcdb_session: Session
+):
+
+    # Check that values are in the DB.
+    desired_qc_state_dict = qcdb_session.execute(
+        select(QcStateDict.id_qc_state_dict).where(
+            QcStateDict.state == qc_status_post.qc_state
+        )
+    ).one_or_none()
+    if desired_qc_state_dict is None:
+        raise Exception(
+            "Desired QC state is not in the QC database. It might not be allowed."
+        )
+
+    user = qcdb_session.execute(
+        select(User).where(User.username == qc_status_post.user)
+    ).scalar_one_or_none()
+    if user is None:
+        raise Exception(
+            "User has not been found in the QC database. Has it been registered?"
+        )
+
+    qc_type = qcdb_session.execute(
+        select(QcType.id_qc_type).where(QcType.qc_type == qc_status_post.qc_type)
+    ).one_or_none()
+    if qc_type is None:
+        raise Exception("QC type is not in the QC database.")
+
+    qc_state_db.user = user
+    qc_state_db.date_updated = datetime.now()
+    qc_state_db.id_qc_state_dict = desired_qc_state_dict[0]
+    qc_state_db.id_qc_type = qc_type[0]
+    qc_state_db.created_by = "LangQC"
+    qc_state_db.is_preliminary = qc_status_post.is_preliminary
+
+
+def qc_status_json(db_qc_state: QcState) -> QcStatus:
+
+    return QcStatus(
+        user=db_qc_state.user.username,
+        date_created=db_qc_state.date_created,
+        date_updated=db_qc_state.date_updated,
+        qc_type=db_qc_state.qc_type.qc_type,
+        qc_type_description=db_qc_state.qc_type.description,
+        qc_state=db_qc_state.qc_state_dict.state,
+        is_preliminary=db_qc_state.is_preliminary,
+        created_by=db_qc_state.created_by,
+    )
